Move around doDelete and Utils
